+++ sickrage/core/webserver/views.py
@@ -55,7 +55,7 @@ from sickrage.core.exceptions import CantRefreshShowException, \
 from sickrage.core.helpers import argToBool, backupSR, check_url, \
     chmodAsParent, findCertainShow, generateApiKey, getDiskSpaceUsage, makeDir, readFileBuffered, \
     remove_article, restoreConfigZip, \
-    sanitizeFileName, tryInt, clean_url
+    sanitizeFileName, tryInt, clean_url, try_int
 from sickrage.core.helpers.browser import foldersAtPath
 from sickrage.core.helpers.compat import cmp
 from sickrage.core.imdb_popular import imdbPopular
@@ -826,6 +826,42 @@ class Home(WebHandler):
             return "Tweet successful, check your twitter to make sure it worked"
         else:
             return "Error sending tweet"
+
+    @staticmethod
+    def testTwilio(account_sid=None, auth_token=None, phone_sid=None, to_number=None):
+        if not sickrage.srCore.notifiersDict['twilio'].account_regex.match(account_sid):
+            return 'Please enter a valid account sid'
+
+        if not sickrage.srCore.notifiersDict['twilio'].auth_regex.match(auth_token):
+            return 'Please enter a valid auth token'
+
+        if not sickrage.srCore.notifiersDict['twilio'].phone_regex.match(phone_sid):
+            return 'Please enter a valid phone sid'
+
+        if not sickrage.srCore.notifiersDict['twilio'].number_regex.match(to_number):
+            return 'Please format the phone number as "+1-###-###-####"'
+
+        result = sickrage.srCore.notifiersDict['twilio'].test_notify()
+        if result:
+            return 'Authorization successful and number ownership verified'
+        else:
+            return 'Error sending sms'
+
+    @staticmethod
+    def testSlack():
+        result = sickrage.srCore.notifiersDict['slack'].test_notify()
+        if result:
+            return "Slack message successful"
+        else:
+            return "Slack message failed"
+
+    @staticmethod
+    def testDiscord():
+        result = sickrage.srCore.notifiersDict['discord'].test_notify()
+        if result:
+            return "Discord message successful"
+        else:
+            return "Discord message failed"
 
     @staticmethod
     def testKODI(host=None, username=None, password=None):
@@ -3654,6 +3690,11 @@ class Config(WebHandler):
             action='index'
         )
 
+    def reset(self):
+        sickrage.srCore.srConfig.load(True)
+        sickrage.srCore.srNotifications.message('Configuration Reset to Defaults', os.path.join(sickrage.CONFIG_FILE))
+        return self.redirect("/config/general")
+
 
 @Route('/config/general(/?.*)')
 class ConfigGeneral(Config):
@@ -3763,11 +3804,10 @@ class ConfigGeneral(Config):
         sickrage.srCore.srConfig.NO_RESTART = sickrage.srCore.srConfig.checkbox_to_value(no_restart)
 
         sickrage.srCore.srConfig.SSL_VERIFY = sickrage.srCore.srConfig.checkbox_to_value(ssl_verify)
-        sickrage.srCore.srConfig.COMING_EPS_MISSED_RANGE = sickrage.srCore.srConfig.to_int(coming_eps_missed_range,
-                                                                                           default=7)
+        sickrage.srCore.srConfig.COMING_EPS_MISSED_RANGE = try_int(coming_eps_missed_range, 7)
         sickrage.srCore.srConfig.DISPLAY_ALL_SEASONS = sickrage.srCore.srConfig.checkbox_to_value(display_all_seasons)
 
-        sickrage.srCore.srConfig.WEB_PORT = sickrage.srCore.srConfig.to_int(web_port)
+        sickrage.srCore.srConfig.WEB_PORT = try_int(web_port)
         sickrage.srCore.srConfig.WEB_IPV6 = sickrage.srCore.srConfig.checkbox_to_value(web_ipv6)
         if sickrage.srCore.srConfig.checkbox_to_value(encryption_version) == 1:
             sickrage.srCore.srConfig.ENCRYPTION_VERSION = 2
@@ -3784,10 +3824,10 @@ class ConfigGeneral(Config):
             sickrage.srCore.srConfig.DATE_PRESET = date_preset
 
         if indexer_default:
-            sickrage.srCore.srConfig.INDEXER_DEFAULT = sickrage.srCore.srConfig.to_int(indexer_default)
+            sickrage.srCore.srConfig.INDEXER_DEFAULT = try_int(indexer_default)
 
         if indexer_timeout:
-            sickrage.srCore.srConfig.INDEXER_TIMEOUT = sickrage.srCore.srConfig.to_int(indexer_timeout)
+            sickrage.srCore.srConfig.INDEXER_TIMEOUT = try_int(indexer_timeout)
 
         if time_preset:
             sickrage.srCore.srConfig.TIME_PRESET_W_SECONDS = time_preset
@@ -3926,7 +3966,7 @@ class ConfigSearch(Config):
 
         sickrage.srCore.srConfig.NZB_METHOD = nzb_method
         sickrage.srCore.srConfig.TORRENT_METHOD = torrent_method
-        sickrage.srCore.srConfig.USENET_RETENTION = sickrage.srCore.srConfig.to_int(usenet_retention, default=500)
+        sickrage.srCore.srConfig.USENET_RETENTION = try_int(usenet_retention, 500)
 
         sickrage.srCore.srConfig.TORRENT_TRACKERS = torrent_trackers if torrent_trackers else ""
         sickrage.srCore.srConfig.IGNORE_WORDS = ignore_words if ignore_words else ""
@@ -3964,7 +4004,7 @@ class ConfigSearch(Config):
         sickrage.srCore.srConfig.NZBGET_CATEGORY_ANIME_BACKLOG = nzbget_category_anime_backlog
         sickrage.srCore.srConfig.NZBGET_HOST = sickrage.srCore.srConfig.clean_host(nzbget_host)
         sickrage.srCore.srConfig.NZBGET_USE_HTTPS = sickrage.srCore.srConfig.checkbox_to_value(nzbget_use_https)
-        sickrage.srCore.srConfig.NZBGET_PRIORITY = sickrage.srCore.srConfig.to_int(nzbget_priority, default=100)
+        sickrage.srCore.srConfig.NZBGET_PRIORITY = try_int(nzbget_priority, 100)
 
         sickrage.srCore.srConfig.TORRENT_USERNAME = torrent_username
         sickrage.srCore.srConfig.TORRENT_PASSWORD = torrent_password
@@ -4294,7 +4334,7 @@ class ConfigProviders(Config):
             sickrage.srCore.providersDict.provider_order += [curProvider]
             if curProvider in sickrage.srCore.providersDict.all():
                 curProvObj = sickrage.srCore.providersDict.all()[curProvider]
-                curProvObj.enabled = bool(sickrage.srCore.srConfig.to_int(curEnabled))
+                curProvObj.enabled = bool(try_int(curEnabled))
 
         # dynamically load provider settings
         for providerID, providerObj in sickrage.srCore.providersDict.all().items():
@@ -4432,6 +4472,9 @@ class ConfigNotifications(Config):
                           prowl_notify_onsubtitledownload=None, prowl_api=None, prowl_priority=0,
                           use_twitter=None, twitter_notify_onsnatch=None, twitter_notify_ondownload=None,
                           twitter_notify_onsubtitledownload=None, twitter_usedm=None, twitter_dmto=None,
+                          use_twilio=None, twilio_notify_onsnatch=None, twilio_notify_ondownload=None,
+                          twilio_notify_onsubtitledownload=None, twilio_phone_sid=None, twilio_account_sid=None,
+                          twilio_auth_token=None, twilio_to_number=None,
                           use_boxcar2=None, boxcar2_notify_onsnatch=None, boxcar2_notify_ondownload=None,
                           boxcar2_notify_onsubtitledownload=None, boxcar2_accesstoken=None,
                           use_pushover=None, pushover_notify_onsnatch=None, pushover_notify_ondownload=None,
@@ -4459,7 +4502,13 @@ class ConfigNotifications(Config):
                           pushbullet_notify_onsubtitledownload=None, pushbullet_api=None, pushbullet_device_list=None,
                           use_email=None, email_notify_onsnatch=None, email_notify_ondownload=None,
                           email_notify_onsubtitledownload=None, email_host=None, email_port=25, email_from=None,
-                          email_tls=None, email_user=None, email_password=None, email_list=None, **kwargs):
+                          email_tls=None, email_user=None, email_password=None, email_list=None, use_slack=None,
+                          slack_notify_onsnatch=None, slack_notify_ondownload=None,
+                          slack_notify_onsubtitledownload=None, slack_webhook=None, use_discord=False,
+                          discord_notify_onsnatch=None, discord_notify_ondownload=None,
+                          discord_notify_onsubtitledownload=None, discord_webhook=None, discord_name=None,
+                          discord_avatar_url=None, discord_tts=None,
+                          **kwargs):
 
         results = []
 
@@ -4548,6 +4597,39 @@ class ConfigNotifications(Config):
         sickrage.srCore.srConfig.TWITTER_USEDM = sickrage.srCore.srConfig.checkbox_to_value(twitter_usedm)
         sickrage.srCore.srConfig.TWITTER_DMTO = twitter_dmto
 
+        sickrage.srCore.srConfig.USE_TWILIO = sickrage.srCore.srConfig.checkbox_to_value(use_twilio)
+        sickrage.srCore.srConfig.TWILIO_NOTIFY_ONSNATCH = sickrage.srCore.srConfig.checkbox_to_value(
+            twilio_notify_onsnatch)
+        sickrage.srCore.srConfig.TWILIO_NOTIFY_ONDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            twilio_notify_ondownload)
+        sickrage.srCore.srConfig.TWILIO_NOTIFY_ONSUBTITLEDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            twilio_notify_onsubtitledownload)
+        sickrage.srCore.srConfig.TWILIO_PHONE_SID = twilio_phone_sid
+        sickrage.srCore.srConfig.TWILIO_ACCOUNT_SID = twilio_account_sid
+        sickrage.srCore.srConfig.TWILIO_AUTH_TOKEN = twilio_auth_token
+        sickrage.srCore.srConfig.TWILIO_TO_NUMBER = twilio_to_number
+
+        sickrage.srCore.srConfig.USE_SLACK = sickrage.srCore.srConfig.checkbox_to_value(use_slack)
+        sickrage.srCore.srConfig.SLACK_NOTIFY_ONSNATCH = sickrage.srCore.srConfig.checkbox_to_value(
+            slack_notify_onsnatch)
+        sickrage.srCore.srConfig.SLACK_NOTIFY_ONDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            slack_notify_ondownload)
+        sickrage.srCore.srConfig.SLACK_NOTIFY_ONSUBTITLEDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            slack_notify_onsubtitledownload)
+        sickrage.srCore.srConfig.SLACK_WEBHOOK = slack_webhook
+
+        sickrage.srCore.srConfig.USE_DISCORD = sickrage.srCore.srConfig.checkbox_to_value(use_discord)
+        sickrage.srCore.srConfig.DISCORD_NOTIFY_ONSNATCH = sickrage.srCore.srConfig.checkbox_to_value(
+            discord_notify_onsnatch)
+        sickrage.srCore.srConfig.DISCORD_NOTIFY_ONDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            discord_notify_ondownload)
+        sickrage.srCore.srConfig.DISCORD_NOTIFY_ONSUBTITLEDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
+            discord_notify_onsubtitledownload)
+        sickrage.srCore.srConfig.DISCORD_WEBHOOK = discord_webhook
+        sickrage.srCore.srConfig.DISCORD_NAME = discord_name
+        sickrage.srCore.srConfig.DISCORD_AVATAR_URL = discord_avatar_url
+        sickrage.srCore.srConfig.DISCORD_TTS = discord_tts
+
         sickrage.srCore.srConfig.USE_BOXCAR2 = sickrage.srCore.srConfig.checkbox_to_value(use_boxcar2)
         sickrage.srCore.srConfig.BOXCAR2_NOTIFY_ONSNATCH = sickrage.srCore.srConfig.checkbox_to_value(
             boxcar2_notify_onsnatch)
@@ -4624,7 +4706,7 @@ class ConfigNotifications(Config):
         sickrage.srCore.srConfig.EMAIL_NOTIFY_ONSUBTITLEDOWNLOAD = sickrage.srCore.srConfig.checkbox_to_value(
             email_notify_onsubtitledownload)
         sickrage.srCore.srConfig.EMAIL_HOST = sickrage.srCore.srConfig.clean_host(email_host)
-        sickrage.srCore.srConfig.EMAIL_PORT = sickrage.srCore.srConfig.to_int(email_port, default=25)
+        sickrage.srCore.srConfig.EMAIL_PORT = try_int(email_port, 25)
         sickrage.srCore.srConfig.EMAIL_FROM = email_from
         sickrage.srCore.srConfig.EMAIL_TLS = sickrage.srCore.srConfig.checkbox_to_value(email_tls)
         sickrage.srCore.srConfig.EMAIL_USER = email_user